requirements.txt
ipykernel
pandas
numpy
matplotlib
plotly
seaborn
scipy
scikit-learn
imblearn
xgboost
catboost
pymongo
from_root
evidently==0.2.8
dill
PyYAML
boto3
mypy-boto3-s3
botocore
fastapi
uvicorn
jinja2
python-multipart
azure-storage-blob
-e .